Refactor native utils
