+++ native/src/main/java/org/ballerinalang/redis/actions/ConnectionActions.java
@@ -25,7 +25,7 @@ import io.ballerina.runtime.api.values.BString;
 import org.ballerinalang.redis.RedisDataSource;
 import org.ballerinalang.redis.utils.ModuleUtils;
 
-import static org.ballerinalang.redis.Constants.ERROR;
+import static org.ballerinalang.redis.Constants.REDIS_ERROR_TYPE;
 
 /**
  * Redis connection actions.
@@ -42,8 +42,8 @@ public class ConnectionActions extends AbstractRedisAction {
         try {
             return StringUtils.fromString(ping(redisDataSource));
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
-                    StringUtils.fromString(e.getMessage()), 
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
+                    StringUtils.fromString(e.getMessage()),
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
     }
@@ -60,8 +60,8 @@ public class ConnectionActions extends AbstractRedisAction {
         try {
             return StringUtils.fromString(auth(password.getValue(), redisDataSource));
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
-                    StringUtils.fromString(e.getMessage()), 
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
+                    StringUtils.fromString(e.getMessage()),
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
     }
@@ -78,9 +78,19 @@ public class ConnectionActions extends AbstractRedisAction {
         try {
             return StringUtils.fromString(echo(message.getValue(), redisDataSource));
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
-                    StringUtils.fromString(e.getMessage()), 
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
+                    StringUtils.fromString(e.getMessage()),
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
+    }
+
+    /**
+     * Close a redis server connection.
+     *
+     * @param redisClient Client from the Ballerina redis client
+     */
+    public static void close(BObject redisClient) {
+        RedisDataSource redisDataSource = (RedisDataSource) redisClient.getNativeData("DATA_SOURCE");
+        close(redisDataSource);
     }
 }
