+++ native/src/main/java/org/ballerinalang/redis/actions/HashActions.java
@@ -27,7 +27,9 @@ import io.ballerina.runtime.api.values.BString;
 import org.ballerinalang.redis.RedisDataSource;
 import org.ballerinalang.redis.utils.ModuleUtils;
 
-import static org.ballerinalang.redis.Constants.ERROR;
+import static org.ballerinalang.redis.Constants.REDIS_ERROR_TYPE;
+import static org.ballerinalang.redis.utils.ConversionUtils.createMapFromBMap;
+import static org.ballerinalang.redis.utils.ConversionUtils.createStringArrayFromBArray;
 
 /**
  * Redis hash actions.
@@ -47,7 +49,7 @@ public class HashActions extends AbstractRedisAction {
         try {
             return hDel(StringUtils.fromString(key.getValue()), redisDataSource, createStringArrayFromBArray(fields));
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -67,7 +69,7 @@ public class HashActions extends AbstractRedisAction {
         try {
             return hExists(key.getValue(), field.getValue(), redisDataSource);
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -86,7 +88,7 @@ public class HashActions extends AbstractRedisAction {
         try {
             return StringUtils.fromString(hGet(key.getValue(), field.getValue(), redisDataSource));
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -104,7 +106,7 @@ public class HashActions extends AbstractRedisAction {
         try {
             return hGetAll(key.getValue(), redisDataSource);
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -124,7 +126,7 @@ public class HashActions extends AbstractRedisAction {
         try {
             return hIncrBy(key.getValue(), field.getValue(), amount, redisDataSource);
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -144,7 +146,7 @@ public class HashActions extends AbstractRedisAction {
         try {
             return hIncrByFloat(key.getValue(), field.getValue(), amount, redisDataSource);
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -162,7 +164,7 @@ public class HashActions extends AbstractRedisAction {
         try {
             return hKeys(key.getValue(), redisDataSource);
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -180,7 +182,7 @@ public class HashActions extends AbstractRedisAction {
         try {
             return hLen(key.getValue(), redisDataSource);
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -199,7 +201,7 @@ public class HashActions extends AbstractRedisAction {
         try {
             return hMGet(key.getValue(), redisDataSource, createStringArrayFromBArray(fields));
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -218,7 +220,7 @@ public class HashActions extends AbstractRedisAction {
         try {
             return StringUtils.fromString(hMSet(key.getValue(), createMapFromBMap(fieldValueMap), redisDataSource));
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -239,7 +241,7 @@ public class HashActions extends AbstractRedisAction {
         try {
             return hSet(key.getValue(), field.getValue(), value.getValue(), redisDataSource);
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -260,7 +262,7 @@ public class HashActions extends AbstractRedisAction {
         try {
             return hSetNx(key.getValue(), field.getValue(), value.getValue(), redisDataSource);
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -279,7 +281,7 @@ public class HashActions extends AbstractRedisAction {
         try {
             return hStrln(key.getValue(), field.getValue(), redisDataSource);
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -297,7 +299,7 @@ public class HashActions extends AbstractRedisAction {
         try {
             return hVals(key.getValue(), redisDataSource);
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
