+++ native/src/main/java/org/ballerinalang/redis/actions/ListActions.java
@@ -26,7 +26,8 @@ import io.ballerina.runtime.api.values.BString;
 import org.ballerinalang.redis.RedisDataSource;
 import org.ballerinalang.redis.utils.ModuleUtils;
 
-import static org.ballerinalang.redis.Constants.ERROR;
+import static org.ballerinalang.redis.Constants.REDIS_ERROR_TYPE;
+import static org.ballerinalang.redis.utils.ConversionUtils.createStringArrayFromBArray;
 
 /**
  * Redis list actions.
@@ -46,7 +47,7 @@ public class ListActions extends AbstractRedisAction {
             RedisDataSource redisDataSource = (RedisDataSource) redisClient.getNativeData("DATA_SOURCE");
             return lPushX(redisKey.getValue(), redisDataSource, createStringArrayFromBArray(values));
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -67,7 +68,7 @@ public class ListActions extends AbstractRedisAction {
             RedisDataSource redisDataSource = (RedisDataSource) redisClient.getNativeData("DATA_SOURCE");
             return bLPop(timeOut, redisDataSource, createStringArrayFromBArray(keys));
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -88,7 +89,7 @@ public class ListActions extends AbstractRedisAction {
             RedisDataSource redisDataSource = (RedisDataSource) redisClient.getNativeData("DATA_SOURCE");
             return bRPop(timeOut, redisDataSource, createStringArrayFromBArray(keys));
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -107,7 +108,7 @@ public class ListActions extends AbstractRedisAction {
             RedisDataSource redisDataSource = (RedisDataSource) redisClient.getNativeData("DATA_SOURCE");
             return lPush(redisKey.getValue(), redisDataSource, createStringArrayFromBArray(values));
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -125,7 +126,7 @@ public class ListActions extends AbstractRedisAction {
             RedisDataSource redisDataSource = (RedisDataSource) redisClient.getNativeData("DATA_SOURCE");
             return StringUtils.fromString(lPop(redisKey.getValue(), redisDataSource));
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -144,7 +145,7 @@ public class ListActions extends AbstractRedisAction {
             RedisDataSource redisDataSource = (RedisDataSource) redisClient.getNativeData("DATA_SOURCE");
             return StringUtils.fromString(lIndex(redisKey.getValue(), index, redisDataSource));
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -167,7 +168,7 @@ public class ListActions extends AbstractRedisAction {
             RedisDataSource redisDataSource = (RedisDataSource) redisClient.getNativeData("DATA_SOURCE");
             return lInsert(key.getValue(), before, pivot.getValue(), value.getValue(), redisDataSource);
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -185,7 +186,7 @@ public class ListActions extends AbstractRedisAction {
             RedisDataSource redisDataSource = (RedisDataSource) redisClient.getNativeData("DATA_SOURCE");
             return lLen(redisKey.getValue(), redisDataSource);
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -205,7 +206,7 @@ public class ListActions extends AbstractRedisAction {
             RedisDataSource redisDataSource = (RedisDataSource) redisClient.getNativeData("DATA_SOURCE");
             return lRange(redisKey.getValue(), startPos, stopPos, redisDataSource);
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -225,7 +226,7 @@ public class ListActions extends AbstractRedisAction {
             RedisDataSource redisDataSource = (RedisDataSource) redisClient.getNativeData("DATA_SOURCE");
             return lRem(redisKey.getValue(), count, value.getValue(), redisDataSource);
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -245,7 +246,7 @@ public class ListActions extends AbstractRedisAction {
             RedisDataSource redisDataSource = (RedisDataSource) redisClient.getNativeData("DATA_SOURCE");
             return StringUtils.fromString(lSet(redisKey.getValue(), index, value.toString(), redisDataSource));
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -265,7 +266,7 @@ public class ListActions extends AbstractRedisAction {
             RedisDataSource redisDataSource = (RedisDataSource) redisClient.getNativeData("DATA_SOURCE");
             return StringUtils.fromString(lTrim(redisKey.getValue(), startPos, startPos, redisDataSource));
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -283,7 +284,7 @@ public class ListActions extends AbstractRedisAction {
             RedisDataSource redisDataSource = (RedisDataSource) redisClient.getNativeData("DATA_SOURCE");
             return StringUtils.fromString(rPop(redisKey.getValue(), redisDataSource));
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -302,7 +303,7 @@ public class ListActions extends AbstractRedisAction {
             RedisDataSource redisDataSource = (RedisDataSource) redisClient.getNativeData("DATA_SOURCE");
             return StringUtils.fromString(rPopLPush(src.getValue(), destination.getValue(), redisDataSource));
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -321,7 +322,7 @@ public class ListActions extends AbstractRedisAction {
             RedisDataSource redisDataSource = (RedisDataSource) redisClient.getNativeData("DATA_SOURCE");
             return rPush(key.getValue(), redisDataSource, createStringArrayFromBArray(values));
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
@@ -340,7 +341,7 @@ public class ListActions extends AbstractRedisAction {
             RedisDataSource redisDataSource = (RedisDataSource) redisClient.getNativeData("DATA_SOURCE");
             return rPush(key.getValue(), redisDataSource, createStringArrayFromBArray(values));
         } catch (Throwable e) {
-            return ErrorCreator.createError(ModuleUtils.getModule(), ERROR,
+            return ErrorCreator.createError(ModuleUtils.getModule(), REDIS_ERROR_TYPE,
                     StringUtils.fromString(e.getMessage()), 
                     ErrorCreator.createError(StringUtils.fromString(e.getMessage()), e), null);
         }
