--- native/src/main/java/org/ballerinalang/redis/endpoint/InitRedisClient.java
@@ -1,93 +1,0 @@
-/*
- * Copyright (c) 2020, WSO2 Inc. (http:www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http:www.apache.orglicensesLICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.ballerinalang.redis.endpoint;
-
-import io.ballerina.runtime.api.utils.StringUtils;
-import io.ballerina.runtime.api.values.BMap;
-import io.ballerina.runtime.api.values.BObject;
-import io.ballerina.runtime.api.values.BString;
-import io.lettuce.core.codec.ByteArrayCodec;
-import io.lettuce.core.codec.RedisCodec;
-import io.lettuce.core.codec.StringCodec;
-import io.lettuce.core.codec.Utf8StringCodec;
-import org.ballerinalang.redis.Constants;
-import org.ballerinalang.redis.RedisDataSource;
-
-/**
- * Creates a Redis client.
- *
- * @since 0.5.0
- */
-public class InitRedisClient {
-
-    /**
-     * Initialize a redis client.
-     *
-     * @param config configuration map
-     */
-    public static void initClient(BObject client, BMap config) {
-        BString host = config.getStringValue(StringUtils.fromString(Constants.EndpointConfig.HOST));
-        BString password = config.getStringValue(StringUtils.fromString(Constants.EndpointConfig.PASSWORD));
-        BMap<BString, BString> options = config.getMapValue(StringUtils.fromString(Constants.EndpointConfig.OPTIONS));
-
-        RedisCodec<String, String> codec = retrieveRedisCodec(Constants.Codec.STRING_CODEC.getCodecName());
-        boolean clusteringEnabled = options.getBooleanValue(StringUtils.fromString(
-                Constants.EndpointConfig.CLUSTERING_ENABLED));
-        boolean poolingEnabled = options.getBooleanValue(StringUtils.fromString(
-                Constants.EndpointConfig.POOLING_ENABLED));
-
-        RedisDataSource dataSource = new RedisDataSource<>(codec, clusteringEnabled, poolingEnabled);
-        dataSource.init(host.toString(), password.toString(), options);
-        client.addNativeData("DATA_SOURCE", dataSource);
-    }
-
-    /**
-     * Retrieve redis codec.
-     *
-     * @param codecString codec string
-     * @return redis codec
-     */
-    private static RedisCodec retrieveRedisCodec(String codecString) {
-        Constants.Codec codec = retrieveCodec(codecString);
-        switch (codec) {
-            case BYTE_ARRAY_CODEC:
-                return new ByteArrayCodec();
-            case STRING_CODEC:
-                return new StringCodec();
-            case UTF8_STRING_CODEC:
-                return new Utf8StringCodec();
-            default:
-                throw new UnsupportedOperationException("Support for RedisCodec " + codec + " is not implemented yet");
-        }
-    }
-
-    /**
-     * Retreive a codec name.
-     *
-     * @param codecString codec string
-     * @return codec name
-     */
-    private static Constants.Codec retrieveCodec(String codecString) {
-        try {
-            return Constants.Codec.fromCodecName(codecString);
-        } catch (IllegalArgumentException e) {
-            throw new RuntimeException("Unsupported Codec: " + codecString);
-        }
-    }
-}
